# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentBroker X",
    "description": "\n**AgentBroker X** is an autonomous agent economy network where AI agents\n**discover, negotiate with, contract, supervise, verify, replace, and pay**\nother AI agents - with no human in the loop.\n\nBuilt for **NANDAHack 2026** (MIT Media Lab + HCLTech).\n\nLifecycle: `discover \u2192 negotiate \u2192 contract \u2192 escrow \u2192 delegate \u2192 verify \u2192\nsettle \u2192 reputation`.\n",
    "contact": {
      "name": "Amaan Khan",
      "url": "https://github.com/amaancoderx"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/tasks/create": {
      "post": {
        "tags": [
          "Task Delegation"
        ],
        "summary": "Create Task",
        "operationId": "create_task_tasks_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/assign": {
      "post": {
        "tags": [
          "Task Delegation"
        ],
        "summary": "Assign Task",
        "operationId": "assign_task_tasks_assign_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskAssign"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/start": {
      "post": {
        "tags": [
          "Task Delegation"
        ],
        "summary": "Start Task",
        "operationId": "start_task_tasks_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/submit": {
      "post": {
        "tags": [
          "Task Delegation"
        ],
        "summary": "Submit Task",
        "operationId": "submit_task_tasks_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskSubmit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/complete": {
      "post": {
        "tags": [
          "Task Delegation"
        ],
        "summary": "Complete Task",
        "description": "Verify, settle escrow, and update reputation - the autonomous close-out.\n\nOn a passing verification: release escrow to the provider and credit a\nsuccessful job. On failure: refund the requester and penalise the provider.",
        "operationId": "complete_task_tasks_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}": {
      "get": {
        "tags": [
          "Task Delegation"
        ],
        "summary": "Get Task",
        "operationId": "get_task_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskAction": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          }
        },
        "type": "object",
        "required": [
          "task_id"
        ],
        "title": "TaskAction"
      },
      "TaskAssign": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "assignee_id": {
            "type": "string",
            "title": "Assignee Id"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "assignee_id"
        ],
        "title": "TaskAssign"
      },
      "TaskCreate": {
        "properties": {
          "requester_id": {
            "type": "string",
            "title": "Requester Id"
          },
          "skill": {
            "type": "string",
            "title": "Skill"
          },
          "spec": {
            "additionalProperties": true,
            "type": "object",
            "title": "Spec"
          },
          "acceptance_criteria": {
            "additionalProperties": true,
            "type": "object",
            "title": "Acceptance Criteria"
          },
          "contract_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Id"
          },
          "deadline_seconds": {
            "type": "integer",
            "title": "Deadline Seconds",
            "default": 120
          }
        },
        "type": "object",
        "required": [
          "requester_id",
          "skill"
        ],
        "title": "TaskCreate"
      },
      "TaskOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "contract_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Id"
          },
          "requester_id": {
            "type": "string",
            "title": "Requester Id"
          },
          "assignee_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Id"
          },
          "skill": {
            "type": "string",
            "title": "Skill"
          },
          "spec": {
            "additionalProperties": true,
            "type": "object",
            "title": "Spec"
          },
          "acceptance_criteria": {
            "additionalProperties": true,
            "type": "object",
            "title": "Acceptance Criteria"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "deadline_seconds": {
            "type": "integer",
            "title": "Deadline Seconds"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "contract_id",
          "requester_id",
          "assignee_id",
          "skill",
          "spec",
          "acceptance_criteria",
          "result",
          "status",
          "attempts",
          "deadline_seconds",
          "created_at"
        ],
        "title": "TaskOut"
      },
      "TaskSubmit": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "result"
        ],
        "title": "TaskSubmit"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}